pyproject.toml
[tool.poetry]
name = "out-of-office"
version = "0.1.0"
description = ""
authors = ["Yryskeldi Azamatov <azamatov0202@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
